requirements.txt
PySimpleGUI==3.29.0
PySimpleGUIWeb==0.25.0
pyserial==3.4